Squashed commit of the following:
commit 2572d7f2f3ba3b24f0534ac167348e6504ee9e22
Author: Jose Daniel Mata <jdmata16@gmail.com>
Date:   Mon Aug 12 13:57:00 2024 -0600

    Arreglos

commit 8ad333476ae62b568ce155c691e7ea616a60ea1e
Merge: 9e42e24 993e32f
Author: Jose Daniel Mata <jdmata16@gmail.com>
Date:   Mon Aug 12 13:48:22 2024 -0600

    Merge remote-tracking branch 'origin/Dylan-Rifa' into Correcto

commit 993e32fcee969d9ec0aa03184ce9503757e9e332
Author: DYL4N1616 <dyurenapiedra@gmail.com>
Date:   Mon Aug 12 00:30:15 2024 -0600

    Anadir rifas -CRUD- Modificacion de  proximas rifas

commit 9e42e24bc3422aa95d30ccceb7e5d32b806fc46a
Author: Jose Daniel Mata <jdmata16@gmail.com>
Date:   Thu Aug 8 17:57:01 2024 -0600

    Diseño listo

commit 3ada492e055ef92aa39bf7b0c144bf1931be9cc8
Author: Jose Daniel Mata <jdmata16@gmail.com>
Date:   Thu Aug 8 16:52:57 2024 -0600

    rifasentrie casi listo

commit c15c65cc744183857b09fe690b68fcc700c6e09c
Author: DYL4N1616 <dyurenapiedra@gmail.com>
Date:   Wed Aug 7 21:59:50 2024 -0600

    Rifas Proximas

    N/A

commit 70dd455a080cb3c8fe7985ade7be2a51f72c5eeb
Author: Jose Daniel Mata <jdmata16@gmail.com>
Date:   Wed Aug 7 21:04:15 2024 -0600

    vista de errores

+++ Proyecto_Ciudad_De_Los_Ninos/Proyecto_Ciudad_De_Los_Ninos/Controllers/AsistenciasController.cs
@@ -188,11 +188,30 @@ namespace Proyecto_Ciudad_De_Los_Ninos.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var asistencia = await _context.Asistencia.FindAsync(id);
-            _context.Asistencia.Remove(asistencia);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            try
+            {
+                var asistencia = await _context.Asistencia.FindAsync(id);
+
+                if (asistencia == null)
+                {
+                    TempData["ErrorMessage"] = "El dato no se encontró o ya ha sido eliminado.";
+                    return RedirectToAction(nameof(Index));
+                }
+
+                _context.Asistencia.Remove(asistencia);
+                await _context.SaveChangesAsync();
+
+                TempData["SuccessMessage"] = "La asistencia se eliminó correctamente.";
+                return RedirectToAction(nameof(Index));
+            }
+            catch
+            {
+                TempData["ErrorMessage"] = "Ocurrió un error al intentar eliminar la asistencia. Por favor, inténtelo de nuevo más tarde.";
+                return RedirectToAction(nameof(Index));
+            }
         }
+
+
         private bool AsistenciaExists(int id)
         {
             return _context.Asistencia.Any(e => e.ID == id);
